src/gradio/requirements.txt
fastapi==0.112.4
gradio==4.43.0
gradio_client==1.3.0
itsdangerous==2.2.0
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.16
langchain-openai==0.1.15
langchain-text-splitters==0.2.1
pydantic-settings==2.4.0
PyMuPDF==1.24.7
python-keycloak==4.3.0
transformers==4.42.3
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
uvicorn==0.30.6
